🔧 refactor(bot): ajustar importação da classe ClassBot
Files Changed:
- *crawjud/bots/caixa/emissor.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/calculadoras/tjdft.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/elaw/andamentos.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/elaw/cadastro.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/elaw/complement.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/elaw/download.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/elaw/prazos.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/elaw/provisao.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/elaw/sol_pags.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/esaj/busca_pags.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/esaj/capa.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/esaj/emissao.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/esaj/movimentacao.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/esaj/protocolo.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/projudi/capa.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/projudi/intimacoes.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/projudi/movimentacao.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/projudi/proc_parte.py*: ajustada a importação da classe ClassBot para o novo caminho.
- *crawjud/bots/projudi/protocolo.py*: ajustada a importação da classe ClassBot para o novo caminho.
