+++ crawjud/bots/projudi/proc_parte.py
@@ -17,7 +17,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webelement import WebElement
 from urllib3.exceptions import MaxRetryError  # noqa: F401
 
-from crawjud.common.bot import ClassBot
+from crawjud.bot import ClassBot
 from crawjud.common.exceptions.bot import ExecutionError
